edited toy case for much faster runtime

+++ cluster_classification/custom_types.py
@@ -53,8 +53,8 @@ class FeModel(BaseModel):
     early_stop: EarlyStop
 
 class RRange(BaseModel):
-    start: int
-    stop: int
+    start: float
+    stop: float
     num_steps: int
     scale: str
     direction: Literal["normal", "reversed"]
